Docker images: set pip timeout to 180
Some developers in the local environment were seeing issues
where poetry/pip installs were timing out, blocking
dependency installs.

We only have one helpful error message from a developer and can't
reproduce the issue, but we're tweaking the Pip timeout here
as a "shot in the dark" to try and reduce the probability
that this happens in the future.